Refactor display and sensor management
- Added a new Display class for handling UI elements on the ST7789 display.
- Removed the unused visualization.h file.
- Changed _attemptCount from int to uint8_t in wifi_manager.h for better memory usage.
- Updated _actuators_system.cpp to clarify relay logic and removed unnecessary motor driver pin setup.
- Simplified relay control logic in relay.cpp by removing time-based checks.
- Introduced EEPROM for storing cover state in stepper_motor.cpp.
- Implemented display initialization and loading views in _display_system.cpp and UI(s).cpp.
- Created home view in v.home.cpp and added basic structure for soil moisture and temperature/humidity views.
- Updated main.cpp to improve serial communication speed and removed unnecessary wifi updates.
- Added light detection functionality in light_detection.cpp.
- Enhanced soil moisture data retrieval in soil_moisture.cpp with better logging.
- Refined temperature and humidity data handling in temp_&_humi.cpp.
- Removed server management code from server_manager.cpp and related files.

+++ include/manager/system/wifi_manager.h
@@ -48,7 +48,7 @@ private:
     const unsigned long TIME_INTERVAL = 1000 * 60 * 3; // 3 hours
     char _cachedTime[6];
     unsigned long _lastAttemptTime;
-    int _attemptCount;
+    uint8_t _attemptCount;
 
     void connect();
     void onConnected();
